Implemented checksum for when hash fails
The checksum is a sum of all the start indexes in highlightArr, since the hash provided by Java does not account for the attributes, go figure. It ensures that if the text was changed while find was hidden to the same text, it will call find again since it would otherwise mess up the highlight array.